Move ConfiguredRuntimeBuilder to its own file (file size discipline)
src/app/runtime/builder.rs grew to 1014 lines after Phase 2 added the
ConfiguredRuntimeBuilder typestate. Extract that struct + its impl into
src/app/runtime/builder_configured.rs. RuntimeBuilder stays in builder.rs.
Public API surface unchanged — re-exported from app::runtime alongside
RuntimeBuilder.

No behavior change.
